system: moved GHCVersion into its own file
